# PrideHub-Backend-main/swagger.yml
openapi: 3.0.0
info:
  title: PrideHub API List
  description: This is a reference api list of PrideHub Backend.
  version: 1.0.0
  contact:
    email: yashmangal240@gmail.com
servers:
  - url: http://localhost:9999/api
    description: Development server
  - url: https://lgbtq-backend.onrender.com/api
    description: Production Server
paths:
...
  /posts:
    get:
      tags:
        - Common
      summary: Get all posts
      parameters:
        - in: query
          name: page
          description: Page Number
        - in: query
          name: limit
          description: Limit of posts per page
        - in: query
          name: sort
          description: Sort by popular (default) or recent
      responses:
        '200':
          description: Posts Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /fetchPosts/{userId}:
    get:
      tags:
        - Common
      summary: Get all posts of a particular user
      parameters:
        - in: path
          name: userId
          description: User ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Posts Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /fetchPost/{postId}:
    get:
      tags:
        - Common
      summary: Get a post
      parameters:
        - in: path
          name: postId
          description: Post ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Post Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /fetchFollowers/{userId}:
    get:
      tags:
        - Common
      summary: Get all followers of a particular user
      parameters:
        - in: path
          name: userId
          description: User ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Followers Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /fetchFollowing/{userId}:
    get:
      tags:
        - Common
      summary: Get all following of a particular user
      parameters:
        - in: path
          name: userId
          description: User ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Following Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /events:
    get:
      tags:
        - Common
      summary: Get all events
      parameters:
        - in: query
          name: page
          description: Page Number
        - in: query
          name: limit
          description: Limit of events per page
        - in: query
          name: sort
          description: Sort by - popular (default) or recent
      responses:
        '200':
          description: Events Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /fetchUser/{userId}:
    get:
      tags:
        - Common
      summary: Get a user
      parameters:
        - in: path
          name: userId
          description: User ID
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: User Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /users:
    get:
      tags:
        - Common
      summary: Get all users
      parameters:
        - in: query
          name: page
          description: Page Number
        - in: query
          name: limit
          description: Limit of users per page
      responses:
        '200':
          description: Users Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /blogs:
    get:
      tags:
        - Common
      summary: Get all blogs
      parameters:
        - in: query
          name: page
          description: Page Number
        - in: query
          name: limit
          description: Limit of blogs per page
        - in: query
          name: sort
          description: Sort by - popular (default) or recent
      responses:
        '200':
          description: Blogs Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /checkToxicity:
    post:
      tags:
        - Common
      summary: Check Toxicity of a text
      requestBody:
        $ref: '#/components/requestBodies/CheckToxicity'
      responses:
        '200':
          description: Toxicity Score Fetched Successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    CheckToxicityInput:
      type: object
      properties:
        text:
          type: string
...
  securitySchemes:
    bearerAuth:
      name: Authorization
      type: http
      scheme: bearer
      in: header
      bearerFormat: JWT
  requestBodies:
...
    CheckToxicity:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CheckToxicityInput'
        description: Check Toxicity
...